🔥 Delete Profile - Abstract 'ProfileDisplay'
